Storefront page (#64)
* Create store item card component

* Add card background and "out of stock" support

* Turn card into link; allow image to be squished on mobile

* Card hover effect

I think that's all

* Create navbar

* Commit the correct index.tsx

* Make "xyz diamonds" its own component

* Back icon; screen-reader accessible diamond icon

* Load balance from user profile

* Make the shadow less harsh in light mode

* Shrink navbar on mobile, per Figma

Otherwise it be squishi

* Storefront hero content

* Attempt making a l o n g rainbow

* A bit of mobile support for the hero

* Move Hero to its own component

* Help modal

* Mobile help modal

Pure CSS pagination!

However, for some reason scrollbars appear in the desktop help modal and idk why

* ayo?? gradient text outline in CSS is possible??

* Add description prop to item cards so they can be used for storefront collection cards as well

* Move back icon into assets/, navbar URLs to config

* Change back link to `history.back()` button

* Make rainbow stretch across page

* Try making the rainbow fit mobile

* Move rainbow to its own component

You know it's bad when you start prefixing class names when you have a library that does that for you

* The tilted rainbow looks fine on mobile now 🎉

* Get collections from store API

* Create Title component

The heading is similar on the leaderboard and storefront pages, and Trevor suggested making a component out of it. Here it is!

https://github.com/acmucsd/membership-portal-ui-v2/pull/42#discussion_r1214094534

* Finish storefront collections view

* Browse all items view

* Clean up files

* Adjust heading sizes

They somewhat match the Figma, but not really. Maybe moving the heading to a separate component was a bad idea if the font sizes are going to be different every time

* Hero animations and interactive decorations 🎥😎

* Require that an item card be EITHER a store item or a collection, per faris feedback

Based on SignInButton

* Address Faris' feedback

* Show costs under items

* Add Imgur etc to image host allowlist (dev only)

`picture` can be null; I decided to update the type instead of blaming the backend data

I also added `screen and` to media queries and updated some rem values per feedback in #59

* Remove <Title> because each use required overriding the font size

* Image load animation

So when it loads, it's not as jarring when it appears

* Remove redundant useEffect

---------

Co-authored-by: Faris Ashai <farisashai@gmail.com>