Update src/pages/admin/store/pickup/index.tsx
Currently the create button is in the middle of the header, which is a bit weird. I think it makes sense to align it on the right with the Future | Past selector

You could do that by removing justify-content: space-between; from .header, then adding some auto margin between the two elements.

Co-authored-by: Sean <ussa.say@gmail.com>